ufmt

Macro uwrite

uwrite!() { /* proc-macro */ }

Write formatted data into a buffer

This macro accepts a format string, a list of arguments, and a ‘writer’. Arguments will be formatted according to the specified format string and the result will be passed to the writer. The writer must have type [&mut] impl uWrite or [&mut] ufmt::Formatter<'_, impl uWrite>. The macro returns the associated Error type of the uWrite-r.

The syntax is similar to core::write! but only a handful of argument types are accepted:

{} - uDisplay

{:?} - uDebug

{:#?} - “pretty” uDebug

Named parameters and “specified” positional parameters ({0}) are not supported.

{{ and }} can be used to escape braces.