Struct ufmt::Formatter

pub struct Formatter<'w, W>where
    W: uWrite + ?Sized,{ /* private fields */ }

Configuration for formatting

Implementations

impl<'w, W> Formatter<'w, W>where W: uWrite + ?Sized,

pub fn debug_list(&mut self) -> Result<DebugList<'_, 'w, W>, W::Error>

Creates a DebugList builder designed to assist with creation of uDebug implementations for list-like structures.

pub fn debug_map(&mut self) -> Result<DebugMap<'_, 'w, W>, W::Error>

Creates a DebugMap builder designed to assist with creation of uDebug implementations for map-like structures.

pub fn debug_set(&mut self) -> Result<DebugSet<'_, 'w, W>, W::Error>

Creates a DebugSet builder designed to assist with creation of uDebug implementations for set-like structures.

pub fn debug_struct( &mut self, name: &str ) -> Result<DebugStruct<'_, 'w, W>, W::Error>

Creates a DebugStruct builder designed to assist with creation of uDebug implementations for structs.

pub fn debug_tuple( &mut self, name: &str ) -> Result<DebugTuple<'_, 'w, W>, W::Error>

Creates a DebugTuple builder designed to assist with creation of uDebug implementations for tuple structs.

impl<'w, W> Formatter<'w, W>where W: uWrite + ?Sized,

pub fn new(writer: &'w mut W) -> Self

Creates a formatter from the given writer

pub fn pretty( &mut self, f: impl FnOnce(&mut Self) -> Result<(), W::Error> ) -> Result<(), W::Error>

Execute the closure with pretty-printing enabled

pub fn write_char(&mut self, c: char) -> Result<(), W::Error>

Writes a character to the underlying buffer contained within this formatter.

pub fn write_str(&mut self, s: &str) -> Result<(), W::Error>

Writes a string slice to the underlying buffer contained within this formatter.